Extract ReregistrationActivator from the monolithic Activator (#563 Sprint 7, A5)
First per-module commit of the activation split: the reregistration schema
(campaigns, audiences junction, submissions + auth_code/magic_token back-fill,
and the audience→junction data migration) moves to a new
FreeFormCertificate\Reregistration\ReregistrationActivator::create_tables(),
matching the existing per-module activator convention (AudienceActivator,
RecruitmentActivator, etc.) and reusing the shared Core\DatabaseHelperTrait.

The core Activator now delegates via the same class_exists guard pattern it
already uses for the other modules. Cross-table concerns (auth_code UNIQUE
upgrade, composite indexes, foreign keys, cron) stay in the orchestrator.
Behavior-preserving — ActivatorTest green; activation order unchanged.
